agents/orchestration/requirements.txt
# Orchestration Agent Dependencies

# Core dependencies
google-adk>=1.0.0
python-dotenv==1.0.0

# Google Cloud dependencies
google-cloud-bigquery>=3.13.0
google-cloud-aiplatform>=1.38.0

# Vertex AI
vertexai>=1.38.0

# Note: Also depends on other agents in the project
# Ensure those agents' dependencies are also installed:
# - agents.schema_mapping
# - agents.validation
# - agents.staging_loader_agent
# - agents.etl_agent

